Architecture: adopt vertical-slice convention formally (#530)
- Add §1.1 to CLAUDE.md documenting the vertical-slice folder pattern,
  cross-feature dependency rules, shared-kernel allow-list, and DI
  extension method convention.
- Introduce VerticalSliceConventionTests: a [Theory] over feature names
  (Analysis, Export) that scans using-directives and fails CI if any
  file reaches into a sibling feature's internal namespace.
- Slim App.axaml.cs (~297 → 51 lines) by extracting DI registrations
  into seven feature extension methods in CAP.Avalonia/DI/:
  AddUpdateFeature, AddAiAssistantFeature, AddExportFeature,
  AddCoreServices, AddCanvasAndPanels, AddSettingsFeature,
  AddPdkOffsetFeature.
- Update ExportViewModelLifetimeTests to scan DI extension files in
  addition to App.axaml.cs (so the singleton-lifetime checks still
  pass after the move).
- Fix repo-root detection in FileSizeLimitTests and the new test to
  handle git worktrees (.git file, not directory).
- Add .github/pull_request_template.md with a vertical-slice checklist.
